rearranged build once more -- now all of NCL is in pkg/src and 'simply gets built' when phylobase is built

+++ src/ncl/nxscharactersblock.h
@@ -926,7 +926,7 @@ class NxsDiscreteDatatypeMapper
 		/* will return NXS_INVALID_STATE_CODE if the char is unknown */
 		int				 GetStateCodeStored(char currChar) const 
 			{
-			return cLookup[currChar];
+			  return cLookup[static_cast<int>(currChar)];
 			}
 		bool			 IsPolymorphic(int stateCode) const;
 		int				 PositionInSymbols(const char currChar) const;
